% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeConnectivityEnrichment.R
\name{computeConnectivityEnrichment}
\alias{computeConnectivityEnrichment}
\title{Compute Connectivity Enrichment}
\usage{
computeConnectivityEnrichment(
  mFC,
  names = NULL,
  mDrugEnrich = SpaceLINCS::mDrugEnrich,
  methods = c("GSEA", "cor"),
  nmin = 15,
  nprune = 250
)
}
\arguments{
\item{mFC}{= A matrix of differential gene expression fold-change of own experiment, the rows name of the matrix must be the genes names}

\item{mDrugEnrich}{= a large matrix represents the gene expression fold change in the presence of different perturbations, either various drugs concentrations or other genetic engineering techniques as gene-knock-in etc..}

\item{methods}{the computation methods of Enrichment, either using GSEA algorithm or the rank correlation.}

\item{nmin}{the minimum number of drugs}

\item{nprune}{takes only the (nprune) top matching drugs for each comparison to reduce the size of the matrix. default is 0 to take the full matrix}

\item{DrugsAnnot}{= is  matrix represents the drugs annotation that contains the targets, mechanism of action , clinical phase, disease area, and indication}

\item{contrast}{is a character string represent the two compared conditions(contrast) as it is provided from the fold-change matrix}
}
\value{
drugs a list of drugs enrichment stats and annotations
}
\description{
Compute Connectivity Enrichment
}
\examples{
# from the data-sets provided as examples within the package load the .rda files
# DrugsAnnot, mDrugEnrich, mFC
drugs <- computeConnectivityEnrichment(
  mFC = mFC, mDrugEnrich = mDrugEnrich, DrugsAnnot = DrugsAnnot,
  methods = "GSEA", nprune = 250
)
}
