% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMoaEnrichment.R
\name{computeMoaEnrichment}
\alias{computeMoaEnrichment}
\title{Get the mechanism of action object}
\usage{
computeMoaEnrichment(res, annot = SpaceLINCS::DrugsAnnot)
}
\arguments{
\item{dsea}{is a drug set enrichment object as an output of the function  getActiveDSEA()}
}
\value{
MoA Mechanism of Action object
}
\description{
Get the mechanism of action object
}
\examples{
# from the data-sets provided as examples within the package load the .rda files
# DrugsAnnot, mDrugEnrich, mFC
res <- getActiveDSEA(mDrugEnrich, DrugsAnnot, mFC, contr = colnames(mFC)[1])
moa <- computeMoaEnrichment(res)
}
