% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDrugConnectivity.R
\name{plotDrugConnectivity}
\alias{plotDrugConnectivity}
\title{Plot Drug Connectivity}
\usage{
plotDrugConnectivity(res, contr, drugs = NULL, nplots = 16)
}
\arguments{
\item{dsea}{is a drug set Enrichment object is the output of getActiveDSEA}
}
\value{
a plot of drug connectivity
}
\description{
Plot Drug Connectivity
}
\examples{
# from the data-sets provided as examples within the package load the .rda files
# DrugsAnnot, mDrugEnrich, mFC
dsea <- getActiveDSEA(mDrugEnrich, DrugsAnnot, mFC, contr = colnames(mFC)[1])
plotDrugConnectivity(dsea)
}
