% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMOA.R
\name{plotMOA}
\alias{plotMOA}
\title{Plot mechanism of action}
\usage{
plotMOA(moa, contr = NULL, type = c("drugClass", "targetGene"), ntop = 20)
}
\arguments{
\item{moa}{is the output object from \code{computeMoaEnrichment()}}

\item{contr}{contrast to select; required when \code{moa} has multiple contrasts}

\item{type}{which MoA enrichment to plot, either "drugClass" or "targetGene"}

\item{ntop}{the number of the top enteries (genes or drug), ntop = 16 as a default value}
}
\value{
plot of mechanism of action
}
\description{
Plot mechanism of action
}
\examples{
# from the data-sets provided as examples within the package load the .rda files
# DrugsAnnot, mDrugEnrich, mFC
res <- computeConnectivityEnrichment(mFC)
moa <- computeMoaEnrichment(res)
plotMOA(moa, contr=1, type="drugClass", ntop=25)
}
