% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDrugConnectivity.R
\name{selectResult}
\alias{selectResult}
\title{Select result}
\usage{
selectResult(res, contr)
}
\arguments{
\item{res}{is the result from computeConnectivityEnrichment()}

\item{contr}{contrast selection}
}
\value{
selected result for selected contrast
}
\description{
Select result
}
\examples{
res <- computeConnectivityEnrichment(mFC)
head(selectResult(res,1))
}
