% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceLINCS.R
\docType{package}
\name{SpaceLINCS}
\alias{SpaceLINCS}
\title{SpaceLINCS: is a new method to visualise correlations between experimental gene expression profiles and drug connectivity map profiles}
\description{
The SpaceLINCS package provides three categories of important functions:
computePeturbEnrichment, computeComboEnrichment and plotDrugConnectivity.
}
\section{SpaceLINCS functions}{


computeConnectivityEnrichment: Compute Connectivity Enrichment  based on database and annotation provided as input parameter
USAGE:
computeConnectivityEnrichment(mFC = mFC, mDrugEnrich = mDrugEnrich, DrugsAnnot = DrugsAnnot, methods = "GSEA", nprune = 250, contrast = NULL)

plotActivationMap: Plot drug Activity Map.
USAGE:
plotActivationMap(res, nterms = 50, nfc = 20)

computeMoaEnrichment: Compute mechanism-of-action enrichment
USAGE:
computeMoaEnrichment(dsea)

plotDrugConnectivity: Plot Drug Connectivity
USAGE:
plotDrugConnectivity(dsea)

plotMOA: Plot mechanism of action bargraph
USAGE:
plotMOA(dsea, ntop = 16)
}

