% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeConnectivityEnrichment.R
\name{computeConnectivityEnrichment}
\alias{computeConnectivityEnrichment}
\title{Compute Connectivity Enrichment}
\usage{
computeConnectivityEnrichment(
  mFC,
  names = NULL,
  mDrugEnrich = metaLINCS::mDrugEnrich,
  nmin = 15,
  nprune = 250
)
}
\arguments{
\item{mFC}{A matrix of differential gene expression fold-change of own experiment, the rows name of the matrix must be the genes names}

\item{names}{Names in case mFC is a vector}

\item{mDrugEnrich}{a large matrix represents the gene expression fold change in the presence of different perturbations, either various drugs concentrations or other genetic engineering techniques as gene-knock-in etc..}

\item{nmin}{the minimum number of experiments in the drug set}

\item{nprune}{takes only the (nprune) top matching drugs for each comparison to reduce the size of the matrix. default is 0 to take the full matrix}
}
\value{
list of drugs enrichment statistics and annotations
}
\description{
Compute Connectivity Enrichment
}
\examples{
# from the data-sets provided as examples within the package load the .rda files
# DrugsAnnot, mDrugEnrich, mFC
res <- computeConnectivityEnrichment(mFC, mDrugEnrich, nmin=15, nprune = 250)
}
