% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMoaEnrichment.R
\name{computeMoaEnrichment}
\alias{computeMoaEnrichment}
\title{Get the mechanism of action object}
\usage{
computeMoaEnrichment(res, annot = SpaceLINCS::DrugsAnnot)
}
\arguments{
\item{res}{is the output object from \code{computeConnectivityEnrichment()}}

\item{annot}{is a drug annotation table with \code{moa} and \code{target} columns}
}
\value{
MoA Mechanism of Action object
}
\description{
Get the mechanism of action object
}
\examples{
# from the data-sets provided as examples within the package load the .rda files
# DrugsAnnot, mDrugEnrich, mFC
res <- computeConnectivityEnrichment(mFC)
moa <- computeMoaEnrichment(res)
}
