% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotActivationMap.R
\name{plotActivationMap}
\alias{plotActivationMap}
\title{Plot drug Activity Map}
\usage{
plotActivationMap(res, nterms = 60, nfc = 20, rot = FALSE)
}
\arguments{
\item{nterms}{integer}

\item{nfc}{integer}

\item{dsea}{drugs set enrichment analysis object as output of getActiveDSEA}

\item{drugs}{a list of drugs enrichment stats and annotations the output of computePeturbEnrichment()}

\item{method}{the computation methods of Enrichment, either using GSEA algorithm or the rank correlation.}

\item{contr}{is a character string represent the two compared conditions(contrast) as it is provided from the fold-change matrix}
}
\value{
plot of drug activity map
}
\description{
Plot drug Activity Map
}
\examples{
# from the data-sets provided as examples within the package load the .rda files
#DrugsAnnot, mDrugEnrich, mFC
res <- computeConnectivityEnrichment(mFC)
plotActivationMap(res, nterms = 50, nfc=20, rot=TRUE)

}
