% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotActivationMap.R
\name{plotActivationMap}
\alias{plotActivationMap}
\title{Plot drug Activity Map}
\usage{
plotActivationMap(res, nterms = 60, nfc = 20, rot = FALSE)
}
\arguments{
\item{res}{is a result object from the output of computeConnectivityEnrichment()}

\item{nterms}{maximum number of terms (integer)}

\item{nfc}{maximum number of contrasts (integer)}

\item{rot}{rotate the figure (boolean)}
}
\value{
plot of drug activity map
}
\description{
Plot drug Activity Map
}
\examples{
# from the data-sets provided as examples within the package load the .rda files
#DrugsAnnot, mDrugEnrich, mFC
res <- computeConnectivityEnrichment(mFC)
plotActivationMap(res, nterms = 50, nfc=20, rot=TRUE)

}
