% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDrugConnectivity.R
\name{plotDrugConnectivity}
\alias{plotDrugConnectivity}
\title{Plot Drug Connectivity}
\usage{
plotDrugConnectivity(res, contr, drugs = NULL, nplots = 25)
}
\arguments{
\item{res}{is a result object from the output of computeConnectivityEnrichment()}

\item{contr}{is a index of the selected contrast}

\item{drugs}{is a vector of selected drugs to show}

\item{nplots}{is the number of plots to show}
}
\value{
a plot of drug connectivity
}
\description{
Plot Drug Connectivity
}
\examples{
# from the data-sets provided as examples within the package load the .rda files
# DrugsAnnot, mDrugEnrich, mFC
res <- computeConnectivityEnrichment(mFC)
plotDrugConnectivity(res, contr=1, nplots=16)
}
