% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMOA.R
\name{plotMOA}
\alias{plotMOA}
\title{Plot mechanism of action}
\usage{
plotMOA(moa, contr = NULL, type = c("drugClass", "targetGene"), ntop = 20)
}
\arguments{
\item{ntop}{the number of the top enteries (genes or drug), ntop = 16 as a default value}

\item{dsea}{is dsea object which is the output of the function getActiveDSEA()}
}
\value{
plot of mechanism of action
}
\description{
Plot mechanism of action
}
\examples{
# from the data-sets provided as examples within the package load the .rda files
# DrugsAnnot, mDrugEnrich, mFC
dsea <- getActiveDSEA(mDrugEnrich, DrugsAnnot, mFC, contr = colnames(mFC)[1])
plotMOA(dsea)
}
